3.7.12, use native Array.prototype.flat() if available
