+++ package.json
@@ -1,11 +1,11 @@
 {
     "author": {
         "name": "Mike Rodrigues",
-        "email": "rodrigues.mi@husky.neu.edu"
+        "email": "rodrigues.mi@northeastern.edu"
     },
     "name": "json-2-csv",
     "description": "A JSON to CSV and CSV to JSON converter that natively supports sub-documents and auto-generates the CSV heading.",
-    "version": "3.7.11",
+    "version": "3.7.12",
     "repository": {
         "type": "git",
         "url": "https://github.com/mrodrig/json-2-csv.git"
@@ -42,12 +42,12 @@
     },
     "devDependencies": {
         "babel-eslint": "10.1.0",
-        "eslint": "7.10.0",
-        "mocha": "8.1.3",
+        "eslint": "7.14.0",
+        "mocha": "8.2.1",
         "nyc": "15.1.0",
         "should": "13.2.3",
         "tslint": "6.1.3",
-        "typescript": "4.0.3"
+        "typescript": "4.1.2"
     },
     "engines": {
         "node": ">= 10"
