+++ test/data/json/nestedFlattening.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const nested = [],
+    expected = [];
+
+for (let a = 0; a < 234567; a++) {
+    nested.push([a]);
+    expected.push(a);
+}
+
+module.exports = {
+    nested,
+    expected
+};
